Create a lib for `mining-sv2-proxy`
